Fix: move test from core to utils
